+++ apps/site/src/app/svgPack.ts
@@ -1,0 +1,4 @@
+import { currentVersion } from './releaseInfo'
+
+export const svgPackHref = '/downloads/uplus-icons.svg.zip'
+export const svgPackDownloadName = `uplus-icons-${currentVersion}.svg.zip`